[Application] Move Env class to Env namespace.
