+++ bin/valkyrja
@@ -13,7 +13,7 @@ declare(strict_types=1);
  */
 
 use Valkyrja\Application\Entry\App;
-use Valkyrja\Application\Env;
+use Valkyrja\Application\Env\Env;
 
 $composerAutoloadFilePossibilities = [
     __DIR__ . '/../../../autoload.php',
