+++ src/Valkyrja/Application/Cli/Command/ClearCacheCommand.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Valkyrja\Application\Cli\Command;
 
-use Valkyrja\Application\Env;
+use Valkyrja\Application\Env\Env;
 use Valkyrja\Cli\Interaction\Factory\Contract\OutputFactory;
 use Valkyrja\Cli\Interaction\Message\Banner;
 use Valkyrja\Cli\Interaction\Message\Message;
