+++ src/Valkyrja/Application/Contract/Application.php
@@ -15,7 +15,7 @@ namespace Valkyrja\Application\Contract;
 
 use Valkyrja\Application\Data\Config;
 use Valkyrja\Application\Data\Data;
-use Valkyrja\Application\Env;
+use Valkyrja\Application\Env\Env;
 use Valkyrja\Application\Support\Component;
 use Valkyrja\Container\Manager\Contract\Container;
