+++ src/Valkyrja/Mail/Provider/ServiceProvider.php
@@ -18,7 +18,7 @@ use Mailgun\HttpClient\HttpClientConfigurator;
 use Mailgun\Mailgun;
 use Override;
 use PHPMailer\PHPMailer\PHPMailer as PHPMailerClient;
-use Valkyrja\Application\Env;
+use Valkyrja\Application\Env\Env;
 use Valkyrja\Container\Manager\Contract\Container;
 use Valkyrja\Container\Provider\Provider;
 use Valkyrja\Log\Logger\Contract\Logger;
