+++ tests/EnvClass.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Valkyrja\Tests;
 
-use Valkyrja\Application\Env as AppEnv;
+use Valkyrja\Application\Env\Env as AppEnv;
 use Valkyrja\Filesystem\Manager\Contract\Filesystem;
 use Valkyrja\Filesystem\Manager\InMemoryFilesystem;
