+++ tests/Unit/Application/ApplicationTest.php
@@ -20,7 +20,7 @@ use Valkyrja\Application\Component;
 use Valkyrja\Application\Contract\Application;
 use Valkyrja\Application\Data\Config;
 use Valkyrja\Application\Data\Data;
-use Valkyrja\Application\Env;
+use Valkyrja\Application\Env\Env;
 use Valkyrja\Application\Exception\RuntimeException;
 use Valkyrja\Application\Valkyrja;
 use Valkyrja\Attribute\Collector\Contract\Collector as AttributeCollectorContract;
