+++ tests/Unit/Filesystem/Provider/ServiceProviderTest.php
@@ -16,7 +16,7 @@ namespace Valkyrja\Tests\Unit\Filesystem\Provider;
 use League\Flysystem\AwsS3V3\AwsS3V3Adapter as FlysystemAwsS3Adapter;
 use League\Flysystem\Local\LocalFilesystemAdapter as FlysystemLocalAdapter;
 use PHPUnit\Framework\MockObject\Exception;
-use Valkyrja\Application\Env;
+use Valkyrja\Application\Env\Env;
 use Valkyrja\Filesystem\Manager\Contract\Filesystem as Contract;
 use Valkyrja\Filesystem\Manager\FlysystemFilesystem;
 use Valkyrja\Filesystem\Manager\InMemoryFilesystem;
